# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Todo API",
    "description": "A complete REST API for managing todos using Go and Gin",
    "version": "1.0"
  },
  "host": "localhost:8082",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
    "/todos": {
      "get": {
        "summary": "Get all todos",
        "description": "Retrieves a list of all todos from the database",
        "tags": [
          "todos"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of todos",
            "type": "array"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new todo",
        "description": "Creates a new todo with the provided title and description",
        "tags": [
          "todos"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "todo",
            "in": "body",
            "required": true,
            "description": "Todo data"
          }
        ],
        "responses": {
          "201": {
            "description": "Todo created successfully"
          },
          "400": {
            "description": "Invalid request body or validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "summary": "Get todo by ID",
        "description": "Retrieves a specific todo by its ID",
        "tags": [
          "todos"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Todo ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo details"
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Todo not found"
          }
        }
      },
      "put": {
        "summary": "Update a todo",
        "description": "Updates an existing todo with the provided ID",
        "tags": [
          "todos"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Todo ID"
          },
          {
            "name": "todo",
            "in": "body",
            "required": true,
            "description": "Updated todo data"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo updated successfully"
          },
          "400": {
            "description": "Invalid ID format or request body"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a todo",
        "description": "Deletes an existing todo by its ID",
        "tags": [
          "todos"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Todo ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo deleted successfully"
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Todo not found"
          }
        }
      }
    }
  }
}